move extract_views_from_urlpatterns to utils
